refactor(runtime): split registry module
