+++ apps/desktop/src-tauri/crates/host-domain/src/lib.rs
@@ -123,27 +123,37 @@ mod tests {
     #[test]
     fn public_api_exports_compile() {
         use super::{
-            AgentSessionDocument, AgentSessionModelSelection, AgentSessionStopRequest,
-            AgentWorkflowState, AgentWorkflows, BeadsCheck, BuildSessionBootstrap, CreateTaskInput,
-            DevServerEvent, DevServerGroupState, DevServerScriptState, DevServerScriptStatus,
-            DevServerTerminalChunk, DirectoryEntry, DirectoryListing, GitBranch,
-            GitCommitAllRequest, GitCommitAllResult, GitConflict, GitConflictAbortRequest,
-            GitConflictAbortResult, GitConflictOperation, GitCurrentBranch, GitDiffScope,
-            GitFileStatusCounts, GitPort, GitPullRequest, GitPullResult, GitPushResult,
-            GitRebaseAbortRequest, GitRebaseAbortResult, GitRebaseBranchRequest,
-            GitRebaseBranchResult, GitResetSnapshot, GitResetWorktreeSelection,
-            GitResetWorktreeSelectionRequest, GitResetWorktreeSelectionResult,
-            GitUpstreamAheadBehind, GitWorktreeStatus, GitWorktreeStatusData,
-            GitWorktreeStatusSnapshot, GitWorktreeStatusSummary, GitWorktreeStatusSummaryData,
-            GitWorktreeSummary, IssueType, PlanSubtaskInput, QaReportDocument, QaVerdict,
-            QaWorkflowVerdict, RepoRuntimeHealthCheck, RepoRuntimeHealthMcp,
-            RepoRuntimeHealthObservation, RepoRuntimeHealthRuntime, RepoRuntimeHealthState,
-            RepoRuntimeMcpStatus, RepoRuntimeStartupFailureKind, RepoRuntimeStartupStage,
-            RepoRuntimeStartupStatus, RepoStoreAttachmentHealth, RepoStoreHealth,
-            RepoStoreHealthCategory, RepoStoreHealthStatus, RepoStoreSharedServerHealth,
-            RepoStoreSharedServerOwnershipState, RuntimeCheck, RuntimeInstanceSummary, RuntimeRole,
-            RuntimeSubagentExecutionMode, SpecDocument, SystemCheck, SystemOpenInToolId,
-            SystemOpenInToolInfo, TaskAction, TaskCard, TaskDirectMergeResult,
+            AgentRuntimeKind, AgentSessionDocument, AgentSessionModelSelection,
+            AgentSessionStopRequest, AgentWorkflowState, AgentWorkflows, BeadsCheck,
+            BuildSessionBootstrap, CreateTaskInput, DevServerEvent, DevServerGroupState,
+            DevServerScriptState, DevServerScriptStatus, DevServerTerminalChunk, DirectoryEntry,
+            DirectoryListing, GitBranch, GitCommitAllRequest, GitCommitAllResult, GitConflict,
+            GitConflictAbortRequest, GitConflictAbortResult, GitConflictOperation,
+            GitCurrentBranch, GitDiffScope, GitFileStatusCounts, GitPort, GitPullRequest,
+            GitPullResult, GitPushResult, GitRebaseAbortRequest, GitRebaseAbortResult,
+            GitRebaseBranchRequest, GitRebaseBranchResult, GitResetSnapshot,
+            GitResetWorktreeSelection, GitResetWorktreeSelectionRequest,
+            GitResetWorktreeSelectionResult, GitUpstreamAheadBehind, GitWorktreeStatus,
+            GitWorktreeStatusData, GitWorktreeStatusSnapshot, GitWorktreeStatusSummary,
+            GitWorktreeStatusSummaryData, GitWorktreeSummary, IssueType, PlanSubtaskInput,
+            QaReportDocument, QaVerdict, QaWorkflowVerdict, RepoRuntimeHealthCheck,
+            RepoRuntimeHealthMcp, RepoRuntimeHealthObservation, RepoRuntimeHealthRuntime,
+            RepoRuntimeHealthState, RepoRuntimeMcpStatus, RepoRuntimeStartupFailureKind,
+            RepoRuntimeStartupStage, RepoRuntimeStartupStatus, RepoStoreAttachmentHealth,
+            RepoStoreHealth, RepoStoreHealthCategory, RepoStoreHealthStatus,
+            RepoStoreSharedServerHealth, RepoStoreSharedServerOwnershipState,
+            RuntimeApprovalCapabilities, RuntimeApprovalReplyOutcome, RuntimeApprovalRequestType,
+            RuntimeCapabilities, RuntimeCheck, RuntimeDefinition, RuntimeDescriptor,
+            RuntimeForkTarget, RuntimeHistoryCapabilities, RuntimeHistoryFidelity,
+            RuntimeHistoryReplay, RuntimeHydratedEventType, RuntimeInstanceSummary,
+            RuntimeOmittedPermissionBehavior, RuntimeOptionalSurfaceCapabilities,
+            RuntimePendingInputVisibility, RuntimePromptInputCapabilities,
+            RuntimePromptInputPartType, RuntimeProvisioningMode, RuntimeQuestionAnswerMode,
+            RuntimeRegistry, RuntimeRole, RuntimeSessionLifecycleCapabilities,
+            RuntimeSessionStartMode, RuntimeStartupReadinessConfig,
+            RuntimeStructuredInputCapabilities, RuntimeSubagentExecutionMode,
+            RuntimeSupportedScope, RuntimeWorkflowCapabilities, SpecDocument, SystemCheck,
+            SystemOpenInToolId, SystemOpenInToolInfo, TaskAction, TaskCard, TaskDirectMergeResult,
             TaskDocumentPresence, TaskDocumentSummary, TaskMetadata, TaskQaDocumentPresence,
             TaskStatus, TaskStore, TaskWorktreeSummary, UpdateTaskPatch, WorkspaceRecord,
         };
@@ -158,6 +168,7 @@ mod tests {
 
         check_types_exported!(
             RuntimeInstanceSummary,
+            AgentRuntimeKind,
             AgentSessionDocument,
             AgentSessionModelSelection,
             AgentSessionStopRequest,
@@ -221,9 +232,34 @@ mod tests {
             DirectoryListing,
             QaVerdict,
             QaWorkflowVerdict,
+            RuntimeApprovalCapabilities,
+            RuntimeApprovalReplyOutcome,
+            RuntimeApprovalRequestType,
+            RuntimeCapabilities,
             RuntimeCheck,
+            RuntimeDefinition,
+            RuntimeDescriptor,
+            RuntimeForkTarget,
+            RuntimeHistoryCapabilities,
+            RuntimeHistoryFidelity,
+            RuntimeHistoryReplay,
+            RuntimeHydratedEventType,
+            RuntimeOmittedPermissionBehavior,
+            RuntimeOptionalSurfaceCapabilities,
+            RuntimePendingInputVisibility,
+            RuntimePromptInputCapabilities,
+            RuntimePromptInputPartType,
+            RuntimeProvisioningMode,
+            RuntimeQuestionAnswerMode,
+            RuntimeRegistry,
             RuntimeRole,
+            RuntimeSessionLifecycleCapabilities,
+            RuntimeSessionStartMode,
+            RuntimeStartupReadinessConfig,
+            RuntimeStructuredInputCapabilities,
             RuntimeSubagentExecutionMode,
+            RuntimeSupportedScope,
+            RuntimeWorkflowCapabilities,
             SpecDocument,
             SystemCheck,
             SystemOpenInToolId,
